Refactor func
